Reorganize the tests. Fix the stub setup issue.
To replace tw_util.get_vc with a stub, we have to add setup and teardown
function for some tests, and adding them requires us to reorganize the tests.
